Move TracingConfigurator
